+++ stream-chat-android-client/api/stream-chat-android-client.api
@@ -1898,22 +1898,14 @@ public abstract interface class io/getstream/chat/android/client/experimental/er
 }
 
 public abstract interface class io/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandler : io/getstream/chat/android/client/experimental/errorhandler/ErrorHandler {
-	public abstract fun onDeleteReactionError (Lio/getstream/chat/android/client/errors/ChatError;Ljava/lang/String;Ljava/lang/String;)Lio/getstream/chat/android/client/utils/Result;
+	public abstract fun onDeleteReactionError (Lio/getstream/chat/android/client/call/Call;Ljava/lang/String;Ljava/lang/String;)Lio/getstream/chat/android/client/call/ReturnOnErrorCall;
 }
 
 public final class io/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandler$DefaultImpls {
 	public static fun compareTo (Lio/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandler;Lio/getstream/chat/android/client/experimental/errorhandler/ErrorHandler;)I
 }
 
-public abstract interface class io/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandlerProposal : io/getstream/chat/android/client/experimental/errorhandler/ErrorHandler {
-	public abstract fun onDeleteReactionError (Lio/getstream/chat/android/client/call/Call;Ljava/lang/String;Ljava/lang/String;)Lio/getstream/chat/android/client/call/ReturnOnErrorCall;
-}
-
-public final class io/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandlerProposal$DefaultImpls {
-	public static fun compareTo (Lio/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandlerProposal;Lio/getstream/chat/android/client/experimental/errorhandler/ErrorHandler;)I
-}
-
-public final class io/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandlerProposalKt {
+public final class io/getstream/chat/android/client/experimental/errorhandler/listeners/DeleteReactionErrorHandlerKt {
 	public static final fun onMessageError (Lio/getstream/chat/android/client/call/Call;Ljava/util/List;Ljava/lang/String;Ljava/lang/String;)Lio/getstream/chat/android/client/call/Call;
 }
