+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/errorhandler/factory/DeleteReactionErrorHandlerFactory.kt
@@ -4,19 +4,19 @@ import io.getstream.chat.android.client.experimental.errorhandler.ErrorHandler
 import io.getstream.chat.android.client.experimental.errorhandler.factory.ErrorHandlerFactory
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
-import io.getstream.chat.android.offline.experimental.errorhandler.listener.DeleteReactionErrorHandlerProposalImpl
+import io.getstream.chat.android.offline.experimental.errorhandler.listener.DeleteReactionErrorHandlerImpl
 import io.getstream.chat.android.offline.experimental.global.GlobalMutableState
 import io.getstream.chat.android.offline.experimental.plugin.logic.LogicRegistry
 import kotlinx.coroutines.CoroutineScope
 
 @ExperimentalStreamChatApi
 /**
- * Factory for [DeleteReactionErrorHandlerProposalImpl]
+ * Factory for [DeleteReactionErrorHandlerImpl]
  */
 public class DeleteReactionErrorHandlerFactory : ErrorHandlerFactory {
 
     override fun create(): ErrorHandler {
-        return DeleteReactionErrorHandlerProposalImpl(
+        return DeleteReactionErrorHandlerImpl(
             scope = CoroutineScope(DispatcherProvider.IO),
             logic = LogicRegistry.get(),
             globalState = GlobalMutableState.getOrCreate(),
